refactor(tool): extract github graphql package
Move the runtime-neutral GitHub GraphQL tool and MCP server out of runtime-codex into a standalone workspace package so other runtimes can reuse it without importing Codex internals.

Closes #213
