+++ packages/cli/tsconfig.json
@@ -12,6 +12,7 @@
       "@gh-symphony/extension-github-workflow": ["../extension-github-workflow/src/index.ts"],
       "@gh-symphony/orchestrator": ["../orchestrator/src/index.ts"],
       "@gh-symphony/runtime-codex": ["../runtime-codex/src/index.ts"],
+      "@gh-symphony/tool-github-graphql": ["../tool-github-graphql/src/index.ts"],
       "@gh-symphony/tracker-file": ["../tracker-file/src/index.ts"],
       "@gh-symphony/tracker-github": ["../tracker-github/src/index.ts"],
       "@gh-symphony/worker": ["../worker/src/index.ts"]
