+++ packages/tool-github-graphql/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./tool.js";
+export * from "./mcp-server.js";
