+++ packages/tool-github-graphql/src/tool.ts
